+++ packages/authhero/src/routes/universal-login/signup.tsx
@@ -6,7 +6,7 @@ import { Bindings, Variables } from "../../types";
 import { fetchVendorSettings, initJSXRoute } from "./common";
 import SignupPage from "../../components/SignUpPage";
 import validatePasswordStrength from "../../utils/password";
-import { getUserByEmailAndProvider } from "../../helpers/users";
+import { getUserByProvider } from "../../helpers/users";
 import { userIdGenerate } from "../../utils/user-id";
 import MessagePage from "../../components/Message";
 import { sendValidateEmailAddress } from "../../emails";
@@ -118,8 +118,8 @@ export const signupRoutes = new OpenAPIHono<{
       ctx.set("client_id", client.id);
       ctx.set("connection", connection);
 
-      const email = loginSession.authParams.username;
-      if (!email) {
+      const { username } = loginSession.authParams;
+      if (!username) {
         throw new HTTPException(400, { message: "Username required" });
       }
 
@@ -164,10 +164,10 @@ export const signupRoutes = new OpenAPIHono<{
         : undefined;
 
       try {
-        const existingUser = await getUserByEmailAndProvider({
+        const existingUser = await getUserByProvider({
           userAdapter: ctx.env.data.users,
           tenant_id: client.tenant.id,
-          email,
+          username: username,
           provider: "auth2",
         });
 
@@ -176,13 +176,13 @@ export const signupRoutes = new OpenAPIHono<{
         }
 
         const email_verified =
-          emailVerificationSession?.authParams.username === email;
+          emailVerificationSession?.authParams.username === username;
 
         const newUser = await getDataAdapter(ctx).users.create(
           client.tenant.id,
           {
             user_id: `auth2|${userIdGenerate()}`,
-            email,
+            email: username,
             email_verified,
             provider: "auth2",
             connection,
@@ -231,7 +231,7 @@ export const signupRoutes = new OpenAPIHono<{
             state={state}
             vendorSettings={vendorSettings}
             error={error.message}
-            email={email}
+            email={username}
           />,
           400,
         );